+++ src/main/java/com/ayuconpon/common/config/JpaAuditingConfig.java
@@ -1,0 +1,8 @@
+package com.ayuconpon.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {}
