+++ src/main/java/com/ayuconpon/common/config/webConfig.java
@@ -1,0 +1,18 @@
+package com.ayuconpon.common.config;
+
+import com.ayuconpon.common.resolver.UserIdArgumentResolver;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
+
+@Configuration
+public class webConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new UserIdArgumentResolver());
+    }
+
+}
