+++ src/main/java/com/ayuconpon/common/exception/BaseCustomException.java
@@ -1,0 +1,11 @@
+package com.ayuconpon.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public abstract class BaseCustomException extends RuntimeException {
+
+    public abstract HttpStatus getStatus();
+
+    public abstract String getMessage();
+
+}
