+++ src/main/java/com/ayuconpon/common/resolver/UserId.java
@@ -1,0 +1,11 @@
+package com.ayuconpon.common.resolver;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.PARAMETER)
+public @interface UserId {
+}
