--- src/main/java/com/ayuconpon/config/JpaAuditingConfig.java
@@ -1,8 +1,0 @@
-package com.ayuconpon.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {}
