--- src/main/java/com/ayuconpon/config/webConfig.java
@@ -1,18 +1,0 @@
-package com.ayuconpon.config;
-
-import com.ayuconpon.resolver.UserIdArgumentResolver;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@Configuration
-public class webConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new UserIdArgumentResolver());
-    }
-
-}
