--- src/main/java/com/ayuconpon/coupon/controller/request/IssueCouponRequest.java
@@ -1,18 +1,0 @@
-package com.ayuconpon.coupon.controller.request;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class IssueCouponRequest {
-
-    @NotNull(message = "쿠폰 발급 요청 아이디가 비어있습니다.")
-    private Long couponId;
-
-    public IssueCouponRequest(Long couponId) {
-        this.couponId = couponId;
-    }
-
-}
