--- src/main/java/com/ayuconpon/coupon/controller/response/IssueCouponResponse.java
@@ -1,16 +1,0 @@
-package com.ayuconpon.coupon.controller.response;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class IssueCouponResponse {
-
-    public Long issuedCouponId;
-
-    public IssueCouponResponse(Long issuedCouponId) {
-        this.issuedCouponId = issuedCouponId;
-    }
-
-}
