--- src/main/java/com/ayuconpon/coupon/domain/value/Status.java
@@ -1,14 +1,0 @@
-package com.ayuconpon.coupon.domain.value;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Status {
-    USED("사용됨"),
-    UNUSED("사용되지 않음");
-
-    private final String status;
-
-}
