--- src/main/java/com/ayuconpon/exception/BaseCustomException.java
@@ -1,11 +1,0 @@
-package com.ayuconpon.exception;
-
-import org.springframework.http.HttpStatus;
-
-public abstract class BaseCustomException extends RuntimeException {
-
-    public abstract HttpStatus getStatus();
-
-    public abstract String getMessage();
-
-}
