--- src/main/java/com/ayuconpon/exception/UserIdFormatException.java
@@ -1,19 +1,0 @@
-package com.ayuconpon.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class UserIdFormatException extends BaseCustomException{
-
-    private static final HttpStatus status = HttpStatus.BAD_REQUEST;
-    private static final String message = "잘못된 형식의 사용자 아이디입니다.";
-
-    @Override
-    public HttpStatus getStatus() {
-        return status;
-    }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-}
