--- src/main/java/com/ayuconpon/resolver/UserId.java
@@ -1,11 +1,0 @@
-package com.ayuconpon.resolver;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface UserId {
-}
