+++ src/main/java/com/ayuconpon/userCoupon/controller/response/IssueUserCouponResponse.java
@@ -1,0 +1,16 @@
+package com.ayuconpon.userCoupon.controller.response;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class IssueUserCouponResponse {
+
+    public Long issuedUserCouponId;
+
+    public IssueUserCouponResponse(Long issuedUserCouponId) {
+        this.issuedUserCouponId = issuedUserCouponId;
+    }
+
+}
